+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# NEXT UP
+# 3.5.1
 
 - Monster feature effects: improve parsing of invisibility to handle situations for things like the Poltergeist which are always invisible, and the Pixie which has Superior Invisibility.
 - Active Aura Effect fixes: Changes in DAE caused Active Aura effects to stop working correctly.